feat(demand): manual demand model 0.1.0 — allocation, OD spread, sources (CALC 0.9.34, EMISSIONS 0.2.0)
docs/2026-09-18_manual_demand_guide.md phase B. The uniform stopgap is replaced: demand.passengers_per_year, group shares and OD weights/pins on the request; groups.py allocates per departure, od_matrix.py spreads over the sellable pairs, sources.py splits every pair by journey length into air / car / induced; places_sold is a float. Summary demand KPIs are the model's own (shift_car_* renamed shift_other_*, placeholder flag false); emission factors re-based on Back-on-Track 2022. Members and the family document carry a demand block (FORMAT 7). Migration 2026-09-19b_manual_demand.sql.

+++ backend/adapters/proposal/gtfs_store.py
@@ -599,7 +599,7 @@ def _build_od_pairs(cur, outbound_trip_id: str, return_trip_id: str) -> list[ODP
             destination_stop_id=r["destination_stop_id"],
             class_main=r["class_main"],
             trip_id=r["trip_id"],
-            places_sold=int(r["places_sold"]),
+            places_sold=float(r["places_sold"]),
             avg_price=float(r["avg_price"]),
         )
         for r in cur.fetchall()
